requirements.txt
Bootstrap-Flask==2.3.0
email-validator==2.0.0.post2
Flask==2.3.2
Flask-Admin==1.6.1
Flask-Login==0.6.2
Flask-Session==0.5.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
Pillow==10.0.0
python-slugify==8.0.1
